feat: add resetNodeVisitOnLoad setting
When set to true (the default and current behavior), the node visit
count gets reset every time a script in the yarn project gets parsed
with that node name. Otherwise, the node visit will not get reset if the
same node name is encountered in subsequent script parsing.

Useful if you want to save/restore visit counts across saves.